+++ client/src/lib.rs
@@ -9,4 +9,5 @@
 //! Phase 6 asks for one written by parties with no common employer or funder; a second language
 //! written by the same hands does not satisfy it and must not be reported as though it did.
 
+pub mod control;
 pub mod secrets;